исправить ошибки ревьера
