+++ src/components/app/app.jsx
@@ -5,14 +5,17 @@ import Header from '../header/header';
 import Loader from '../loader/loader';
 import { getIngredients } from '../../services/actions/ingredients';
 import ModalSwitch from '../modal-switch/modal-switch';
+import { getUserData } from '../../services/actions/user';
 
 const App = () => {
   const ingredientsRequest = useSelector(state => state.ingredientsData.ingredientsRequest);
   const dispatch = useDispatch();
+  const accessToken = useSelector((state) => state.userData.accessToken);
 
   useEffect(() => {
     dispatch(getIngredients());
-  }, [dispatch])
+    accessToken && dispatch(getUserData(accessToken));
+  }, [dispatch, accessToken])
 
   return (
     <div className={`${appStyles.app}`}>
