+++ src/components/header/header.jsx
@@ -1,10 +1,8 @@
 import { NavLink } from 'react-router-dom';
 import { Logo, BurgerIcon, ListIcon, ProfileIcon } from '@ya.praktikum/react-developer-burger-ui-components'
 import headerStyles from './header.module.css';
-import { useSelector } from "react-redux";
 
 const Header = () => {
-  const userData = useSelector((state) => state.userData.userData);
 
   return (
     <header className={`${headerStyles.header} text text_type_main-default pt-4 pb-4`}>
@@ -28,19 +26,10 @@ const Header = () => {
         <div className={headerStyles.logo}>
           <Logo />
         </div>
-        {
-          userData ? (
-            <NavLink activeClassName={headerStyles.link_active} className={`p-5 ${headerStyles.link} ${headerStyles.profile}`} to="/profile">
-              <ProfileIcon type="secondary" />
-              <span className={`ml-2`}>Личный кабинет</span>
-            </NavLink>
-          ) : (
-            <NavLink activeClassName={headerStyles.link_active} className={`p-5 ${headerStyles.link} ${headerStyles.profile}`} exact to="/login">
-              <ProfileIcon type="secondary" />
-              <span className={`ml-2`}>Войти</span>
-            </NavLink>
-          )
-        }
+        <NavLink activeClassName={headerStyles.link_active} className={`p-5 ${headerStyles.link} ${headerStyles.profile}`} to="/profile">
+          <ProfileIcon type="secondary" />
+          <span className={`ml-2`}>Личный кабинет</span>
+        </NavLink>
       </div>
     </header >
   )
