+++ src/components/order-history/order-history.jsx
@@ -1,0 +1,3 @@
+const OrderHistory = () => null;
+
+export default OrderHistory;
