+++ src/components/profile-form/profile-form.module.css
@@ -1,0 +1,11 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.buttons_container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
