+++ src/components/protected-route/protected-route.jsx
@@ -1,25 +1,19 @@
-import { Route, Redirect } from 'react-router-dom';
-import { useEffect } from 'react';
-import { useDispatch, useSelector } from "react-redux";
-import { getUserData } from '../../services/actions/user';
+import { Route, Redirect, useLocation } from 'react-router-dom';
+import { useSelector } from "react-redux";
 import PropTypes from "prop-types";
 
 const ProtectedRoute = ({ children, ...rest }) => {
-  const dispatch = useDispatch();
-  const accessToken = useSelector((state) => state.userData.accessToken);
   const userData = useSelector((state) => state.userData.userData);
-
-  useEffect(() => {
-    if (!userData) {
-      dispatch(getUserData(accessToken));
-    }
-  }, [dispatch, accessToken, userData])
+  const location = useLocation();
 
   return (
     <Route
       {...rest}
       render={
-        () => userData ? (children) : <Redirect to="/login" />
+        () => userData ? (children) : <Redirect to={{
+          pathname: `/login`,
+          state: { previousLocation: location },
+        }} />
       }
     />
   );
