+++ src/pages/not-found/not-found.jsx
@@ -6,20 +6,20 @@ import {
 
 const NotFound = () => {
   return (
-    <section className={NotFoundStyles.wrapper}>
+    <article className={NotFoundStyles.wrapper}>
       <p className="text text_type_digits-large text_color_inactive">404</p>
-      <p className="text text_type_main-large">
+      <h1 className="text text_type_main-large">
         Страница не найдена
-      </p>
-      <p className="text text_type_main-medium">
+      </h1>
+      <h2 className="text text_type_main-medium">
         Ну не найдена и не найдена. У нас есть очень крутые бургеры!
-      </p>
+      </h2>
       <Link className={NotFoundStyles.link} to="/">
         <Button type="primary" size="medium">
           К бургерам
         </Button>
       </Link>
-    </section>
+    </article>
   );
 };
