+++ src/pages/profile/profile.jsx
@@ -1,74 +1,21 @@
-import { useState, useRef, useEffect } from "react";
 import ProfileStyles from "./profile.module.css";
 import { NavLink, Route, useRouteMatch } from "react-router-dom";
-import { Input, Button } from "@ya.praktikum/react-developer-burger-ui-components";
-import { useDispatch, useSelector } from "react-redux";
-import { sendUserData, logout } from "../../services/actions/user";
+import { useDispatch, } from "react-redux";
+import { logout } from "../../services/actions/user";
+import ProfileForm from '../../components/profile-form/profile-form';
+import OrderHistory from '../../components/order-history/order-history';
 
 const Profile = () => {
-  const [nameValue, setNameValue] = useState("?");
-  const [loginValue, setLoginValue] = useState("?");
-  const [passwordValue, setPasswordValue] = useState("");
-  const [isDataChanged, setIsDataChanged] = useState(false);
-  const nameInputRef = useRef(null);
-  const emailInputRef = useRef(null);
-  const passwordInputRef = useRef(null);
   const dispatch = useDispatch();
-  const accessToken = useSelector((state) => state.userData.accessToken);
-  const userData = useSelector((state) => state.userData.userData);
   const { path, url } = useRouteMatch();
-
-  const onNameClick = () => nameInputRef.current.focus();
-
-  const oтEmailClick = () => emailInputRef.current.focus();
-
-  const onPasswordClick = () => passwordInputRef.current.focus();
-
-  const onNameChange = (evt) => {
-    const value = evt.target.value
-    setNameValue(value)
-    value === userData.name ? setIsDataChanged(false) : setIsDataChanged(true)
-  }
-
-  const onEmailChange = (evt) => {
-    const value = evt.target.value
-    setLoginValue(value)
-    value === userData.email ? setIsDataChanged(false) : setIsDataChanged(true)
-  }
-
-  const onPasswordChange = (evt) => {
-    const value = evt.target.value
-    setPasswordValue(value)
-    value === passwordValue ? setIsDataChanged(false) : setIsDataChanged(true)
-  }
-
-  const onSubmit = (evt) => {
-    evt.preventDefault();
-    dispatch(sendUserData(accessToken, nameValue, loginValue, passwordValue))
-  }
-
-  const onCancelEditing = (evt) => {
-    evt.preventDefault();
-    setNameValue(userData.name)
-    setLoginValue(userData.email)
-    setPasswordValue('')
-  }
 
   const handleLogout = () => {
     const refreshToken = localStorage.getItem('refreshToken');
     dispatch(logout(refreshToken))
   }
 
-  useEffect(() => {
-    if (userData) {
-      setLoginValue(userData.email);
-      setNameValue(userData.name);
-      setPasswordValue('');
-    }
-  }, [userData]);
-
   return (
-    <section className={ProfileStyles.wrapper}>
+    <article className={ProfileStyles.wrapper}>
       <nav className={ProfileStyles.navigation}>
         <ul className={`${ProfileStyles.list}`}>
           <li className={ProfileStyles.list_item}>
@@ -108,62 +55,12 @@ const Profile = () => {
         </p>
       </nav>
       <Route exact path={`${path}`}>
-        <form onSubmit={onSubmit} className={ProfileStyles.form}>
-          <Input
-            type={"text"}
-            placeholder={"Имя"}
-            onChange={onNameChange}
-            icon={"EditIcon"}
-            value={nameValue}
-            name={"name"}
-            error={false}
-            ref={nameInputRef}
-            onIconClick={onNameClick}
-            errorText={"Ошибка"}
-            size={"default"}
-          />
-          <Input
-            type={"text"}
-            placeholder={"Логин"}
-            onChange={onEmailChange}
-            icon={"EditIcon"}
-            value={loginValue}
-            name={"name"}
-            error={false}
-            ref={emailInputRef}
-            onIconClick={oтEmailClick}
-            errorText={"Ошибка"}
-            size={"default"}
-          />
-          <Input
-            type={"text"}
-            placeholder={"Пароль"}
-            onChange={onPasswordChange}
-            icon={"EditIcon"}
-            value={passwordValue}
-            name={"name"}
-            error={false}
-            ref={passwordInputRef}
-            onIconClick={onPasswordClick}
-            errorText={"Ошибка"}
-            size={"default"}
-          />
-          {
-            isDataChanged && (<div className={ProfileStyles.buttons_container}>
-              <Button onClick={onCancelEditing} type="secondary" size="medium">
-                Отмена
-              </Button>
-              <Button type="primary" size="medium">
-                Сохранить
-              </Button>
-            </div>)
-          }
-        </form>
+        <ProfileForm />
       </Route>
       <Route exact path={`${path}/orders`}>
-        {null}
+        <OrderHistory />
       </Route>
-    </section>
+    </article>
   );
 };
