+++ src/pages/register/register.jsx
@@ -1,13 +1,13 @@
-import { useState, useRef } from "react";
+import { useState, useRef, useEffect } from "react";
 import RegisterStyles from "./register.module.css";
-import { Link } from "react-router-dom";
+import { Link, useHistory } from "react-router-dom";
 import {
   Input,
   Button,
   PasswordInput,
 } from "@ya.praktikum/react-developer-burger-ui-components";
 import { registration } from "../../services/actions/user";
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 const Register = () => {
   const [nameValue, setNameValue] = useState("");
@@ -15,6 +15,8 @@ const Register = () => {
   const [passwordValue, setPasswordValue] = useState("");
   const inputRef = useRef(null);
   const dispatch = useDispatch();
+  const userData = useSelector((state) => state.userData.userData);
+  const history = useHistory();
 
   const onPasswordChange = (e) => {
     setPasswordValue(e.target.value);
@@ -30,12 +32,16 @@ const Register = () => {
     dispatch(registration(emailValue, nameValue, passwordValue));
   };
 
+  useEffect(() => {
+    userData && history.push('/')
+  }, [userData, history])
+
   return (
-    <section className={RegisterStyles.wrapper}>
+    <article className={RegisterStyles.wrapper}>
       <form onSubmit={handleSubmit} className={RegisterStyles.form}>
-        <p className={`${RegisterStyles.title} text text_type_main-medium`}>
+        <h1 className={`${RegisterStyles.title} text text_type_main-medium`}>
           Регистрация
-        </p>
+        </h1>
         <div className="mt-6 mb-6">
           <Input
             type={"text"}
@@ -79,7 +85,7 @@ const Register = () => {
           Войти
         </Link>
       </p>
-    </section>
+    </article>
   );
 };
